(SIMP-8589) Fix ssldir sudoers with bolt
Fixed:
  - Ensure that the sudoers rule for removing the puppet ssldir is not created
    when running from bolt since the directory target is changed at each bolt
    run and will result in non-idempotency.
  - Un-pinned the firewalld module version in .fixtures.yml because that no
    longer appears to cause issues.
  - Allow the local yum repos to optionally specify gpgkey or baseurl strings
    since, technically, both are optional in the `yumrepo` type if they already
    exist on disk.
Workarounds:
  - Pinned the rspec-expectations gem to `~> 3.9.0` until rspec-puppet
    2.8.0 can be released.

SIMP-8589 #close

+++ manifests/admin.pp
@@ -177,17 +177,21 @@ class simp::admin (
     passwd    => !$passwordless_admin_sudo
   }
 
-  # This logic is to avoid allowing admins to run `rm -rf` when the fact
-  # doesn't exist
-  case $facts['puppet_settings'] {
-    Hash:    { $_ssldir = $facts['puppet_settings']['main']['ssldir'] }
-    default: { $_ssldir = '/etc/puppetlabs/puppet/ssl' }
-  }
-  sudo::user_specification { 'admin clean puppet certs':
-    user_list => ["%${admin_group}"],
-    runas     => 'root',
-    cmnd      => ["/bin/rm -rf ${$_ssldir}"],
-    passwd    => !$passwordless_admin_sudo
+  # Bolt sets this to a random directory every time it runs
+  unless simplib::in_bolt() {
+    # This logic is to avoid allowing admins to run `rm -rf` when the fact
+    # doesn't exist
+    case $facts['puppet_settings'] {
+      Hash:    { $_ssldir = $facts['puppet_settings']['main']['ssldir'] }
+      default: { $_ssldir = '/etc/puppetlabs/puppet/ssl' }
+    }
+
+    sudo::user_specification { 'admin clean puppet certs':
+      user_list => ["%${admin_group}"],
+      runas     => 'root',
+      cmnd      => ["/bin/rm -rf ${$_ssldir}"],
+      passwd    => !$passwordless_admin_sudo
+    }
   }
 
   $_polkit_ensure = ($set_polkit_admin_group and $facts['os']['release']['major'] >= '7') ? {
